// src/app/enum.ts
export enum FileItemStatus {
  IDLE = "IDLE",
  IN_QUEUE = "IN_QUEUE",
  UPLOADING = "UPLOADING",
  UPLOADED = "UPLOADED",
  FAILED = "FAILED",
  CANCELLED = "CANCELLED",
}

export enum FileItemEvents {
  ON_ITEM_UPDATED = "ON_ITEM_UPDATED",
  ON_FILE_PROCESS_START = "ON_FILE_PROCESS_START",
  ON_FILE_PROCESS_PROGRESS = "ON_FILE_PROCESS_PROGRESS",
  ON_FILE_PROCESS_COMPLETE = "ON_FILE_PROCESS_COMPLETE",
  ON_FILE_PROCESS_CANCELLED = "ON_FILE_PROCESS_CANCELLED",
  ON_FILE_PROCESS_FAILED = "ON_FILE_PROCESS_FAILED",
}

export enum StoreActions {}